requirements.txt
aiogram==3.19.0
openpyxl
python-dotenv
SQLAlchemy>=2.0
asyncpg